+++ honestcash-v2/src/app/app.component.spec.ts
@@ -2,7 +2,7 @@ import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
 import { Store } from '@ngrx/store';
-import { StoreMock } from './store/store.mock';
+import { StoreMock } from './core/mocks/store.mock';
 import { AuthService } from './core/services/auth.service';
 
 describe('AppComponent', () => {
