+++ honestcash-v2/src/app/app.module.ts
@@ -8,11 +8,11 @@ import { FormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { StoreModule } from '@ngrx/store';
-import { AuthEffects } from './store/auth/auth.effects';
-import { WalletEffects } from './store/wallet/wallet.effects';
-import { UserEffects } from './store/user/user.effects';
+import { AuthEffects } from './core/store/auth/auth.effects';
+import { WalletEffects } from './core/store/wallet/wallet.effects';
+import { UserEffects } from './core/store/user/user.effects';
 
-import { reducers, metaReducers } from './store/app.states';
+import { reducers, metaReducers } from './app.states';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { environment } from '../environments/environment';
 import { AuthGuardService as AuthGuard } from './core/guards/auth-guard.service';
