+++ honestcash-v2/src/app/core/services/user.service.ts
@@ -6,7 +6,7 @@ import { Router } from '@angular/router';
 import { IAuthRequestSuccessResponse } from './auth.service';
 import { WalletService } from './wallet.service';
 import { Store } from '@ngrx/store';
-import { AppState, selectUserState } from '../../store/app.states';
+import { AppState, selectUserState } from '../../app.states';
 
 const log = new Logger('UserService');
