+++ honestcash-v2/src/app/core/store/user/user.actions.ts
@@ -1,5 +1,5 @@
 import { Action } from '@ngrx/store';
-import { IAuthRequestSuccessResponse } from '../../core/services/auth.service';
+import { IAuthRequestSuccessResponse } from '../../services/auth.service';
 
 export enum UserActionTypes {
   USER_SETUP = '[User] User setup',
