+++ honestcash-v2/src/app/core/store/user/user.effects.ts
@@ -7,8 +7,8 @@ import {
   UserActionTypes,
   UserSetup,
 } from './user.actions';
-import { AuthService, IAuthRequestSuccessResponse } from '../../core/services/auth.service';
-import { UserService } from '../../core/services/user.service';
+import { AuthService, IAuthRequestSuccessResponse } from '../../services/auth.service';
+import { UserService } from '../../services/user.service';
 
 @Injectable()
 export class UserEffects {
