+++ honestcash-v2/src/app/modules/welcome/pages/login/login.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, HostBinding } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { NgForm } from '@angular/forms';
-import { AppState } from '../../../../store/app.states';
-import { LogIn } from '../../../../store/auth/auth.actions';
+import { AppState } from '../../../../app.states';
+import { LogIn } from '../../../../core/store/auth/auth.actions';
 import User from '../../../../models/user';
 
 interface LoginForm extends NgForm {
