+++ honestcash-v2/src/app/modules/welcome/pages/status/status.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, HostBinding } from '@angular/core';
 import { Store } from '@ngrx/store';
 
-import { AppState } from '../../../../store/app.states';
-import { GetStatus } from '../../../../store/auth/auth.actions';
+import { AppState } from '../../../../app.states';
+import { GetStatus } from '../../../../core/store/auth/auth.actions';
 
 @Component({
   selector: 'app-welcome-page-status',
